+++ src/tags/html.ts
@@ -1,0 +1,8 @@
+import { WireNode } from "../api/wire_node";
+
+export class HTML extends WireNode {
+    constructor() {
+        super();
+        this.setupNode();
+    }
+}
